Corrected repair encodings
We were only identifying potential regulators responsible for the positive observations. While that identified a good portion of regulators and managed to find solutions for a lot of cases, by not identifying the potential regulators responsible for negative observations as well, there would be certain cases where no solution would be found, even though one existed.

Solution: We consider every compound again when generating nodes. In terms of performance, there is no noticeable difference, even in the largest instances. This method ensures correctness, while also allowing the encodings to be simplified.

Alternative solution: We would need to add extra rules for unique negative observations, which would also require us to identify each unique negative observation on top of the identification of unique positive observations. The complexity this would bring both to the pre-processing code and the encoding, for what seems to be a negligible improvement in performance, makes this solution less desirable.